Fix circular dependency
by extracting mini-automatch into component
